+++ scripts/generate-api-toc.js
@@ -1,12 +1,14 @@
+"use strict";
 // From https://github.com/hapijs/joi/blob/master/generate-readme-toc.js
 
 // Load modules
-import toc from "markdown-toc";
-import fs from "fs";
 
-const { version } = JSON.parse(fs.readFileSync("package.json"));
+const Toc = require("markdown-toc");
+const Fs = require("fs");
+const { version } = require("../package.json");
 
 // Declare internals
+
 const filenames = getFileNames();
 
 function getFileNames() {
@@ -15,7 +17,7 @@ function getFileNames() {
 }
 
 function generate(filename) {
-  const api = fs.readFileSync(filename, "utf8");
+  const api = Fs.readFileSync(filename, "utf8");
   const tocOptions = {
     bullets: "-",
     slugify: function (text) {
@@ -26,10 +28,10 @@ function generate(filename) {
     },
   };
 
-  const output = toc.insert(api, tocOptions)
+  const output = Toc.insert(api, tocOptions)
     .replace(/<!-- version -->(.|\n)*<!-- versionstop -->/, `<!-- version -->\n# ${version} API Reference\n<!-- versionstop -->`);
 
-  fs.writeFileSync(filename, output);
+  Fs.writeFileSync(filename, output);
 }
 
 filenames.forEach(generate);