+++ test/fetch-test.js
@@ -1,8 +1,10 @@
-import express from "express";
-import nock from "nock";
+"use strict";
 
-import { app } from "../app/app.js";
-import Browser from "../index.js";
+const express = require("express");
+const nock = require("nock");
+
+const { app } = require("../app/app.js");
+const Browser = require("../index.js");
 
 describe("window.fetch", () => {
   beforeEach(nock.cleanAll);
@@ -298,7 +300,7 @@ describe("window.fetch", () => {
       });
 
     await browser.window.fetch("https://blahonga.expressen.se/");
-    expect(cookie).to.eql([ "_ga=1" ]);
+    expect(cookie).to.eql("_ga=1");
   });
 
   it("exposes pendingRequests promise list", async () => {
@@ -416,7 +418,7 @@ describe("window.fetch", () => {
 
         const resp = await browser.window.fetch("/redirect").then((r) => r.json());
 
-        expect(resp).to.eql({ cookie: [ "_ga=2" ] });
+        expect(resp).to.eql({ cookie: "_ga=2" });
       });
     });
 
@@ -513,7 +515,7 @@ describe("window.fetch", () => {
 
       const browser = await new Browser(localApp).navigateTo("/");
       const resp = await browser.window.fetch("/redirect", { redirect: "manual" });
-      expect(resp.headers.get("location")).to.equal("https://www.example.com/");
+      expect(resp.headers.get("location")).to.equal("https://www.example.com");
     });
 
     it("redirect from remote resource to local resource is handled", async () => {